==> code/04_supplementary_analysis.Rmd <==
---
title: "Supplementary analysis"
output:
  html_document:
    df_print: paged
---

+ Refit model with imputation of missing covariates

```{r setup, echo = FALSE}

set.seed(0000)

library(tidyverse)
library(patchwork)
library(brms)
library(sjPlot)
library(broom.mixed)
library(mice)

theme_set(theme_minimal())

figdir <- "figures/fit"

# Fix default ggsave background
ggsave <- function(..., bg = 'white') ggplot2::ggsave(..., bg = bg)

```

```{r load_data, echo = FALSE}

source(here::here("code/load_data.R"), echo=TRUE)
dat_nona <- readRDS(here::here("data","analysis_data_primary.rds"))
fit_base <- readRDS(here::here("output","fit_base.rds"))
fit_list_pov <- readRDS(here::here("output","fit_list_pov.rds"))

cluster_PNRs <- readRDS(here::here("data","cluster_ests_95_prev31_notif25_rintc.rds"))

priors <- c(prior(normal(-2, 2), class = Intercept),
            prior(normal(0,10), class = b),
            prior(cauchy(0,1), class = sd, group = "clustid"))

```

# Adjusting for prior TB prevalence

```{r add_tbprev}

fit_base2 <- brms:::update.brmsfit(fit_base, 
                                   formula = "test01 ~ sex*agegp + hiv + clust_inc_1518 + (1|clustid)",
                                   newdata = dat_nona)

saveRDS(fit_base2, here::here("output","fit_base_clustinc.rds"))

print(summary(fit_base))
print(summary(fit_base2))

print(exp(fixef(fit_base2)))

```

## Testing prevalence versus prevalence-to-notification ratios

```{r comp_pnr}

mods <- list(Baseline = fit_base, `Poverty-adjusted: PMT score` = fit_list_pov[[1]], `Poverty-adjusted: Self-assessed` = fit_list_pov[[2]])

get_iid <- function(x, nm){
  x %>%
    spread_draws(r_clustid[clustid,]) %>%
    summarise_draws(mean, ~quantile(.x, probs = c(0.025, 0.975))) %>% 
    mutate(Model = nm) %>% 
    return()
}

get_pred <- function(x, nm){
  x %>%
    fitted() %>%
    as_tibble() %>% 
    mutate(Model = nm,
           clustid = dat_nona$clustid) %>% 
    return()
}

purrr::imap(mods,get_iid) %>% 
  bind_rows() -> compare_iids

# purrr::imap(mods,get_pred) %>% 
#   bind_rows() -> compare_preds

compare_iids %>% 
  rename(cluster = clustid) %>% 
  full_join(cluster_PNRs) %>% 
  ggplot(aes(prev_to_notif_ratio, mean, 
             xmin = .lower, xmax = .upper,
             ymin = `2.5%`, ymax = `97.5%`)) +
  geom_point(alpha = 0.7) +
  geom_errorbar(alpha = 0.1) +
  geom_errorbarh(alpha = 0.1) +
  facet_wrap(~Model) +
  scale_x_continuous(trans = "log2") +
  labs(x = "Prevalence-to-notification ratio", y = "Testing history model: cluster residual") +
  theme(text = element_text(size = 14)) -> plot_resid_PNR

# compare_preds %>% 
#   rename(cluster = clustid) %>% 
#   full_join(cluster_PNRs) %>%
#   ggplot(aes(prev_to_notif_ratio, Estimate, 
#              xmin = .lower, xmax = .upper,
#              ymin = `Q2.5`, ymax = `Q97.5`)) +
#   geom_point(alpha = 0.2) +
#   # geom_errorbar(alpha = 0.05) +
#   # geom_errorbarh(alpha = 0.05) +
#   facet_wrap(~Model) +
#   scale_x_continuous(trans = "log2") +
#   labs(x = "Prevalence-to-notification ratio", y = "Testing history model: fitted prevalence") -> plot_fit_PNR

plot_resid_PNR

# plot_fit_PNR

ggsave(here::here("figures/manuscript/supplementary","figs9.png"), 
       plot_resid_PNR, height = 4, width = 12)

```

# Testing prevalence versus case notification rates

```{r test_CNR_ratios}

cnrs20 %>% 
  mutate(year = lubridate::year(datereg),
         clustid = as.numeric(gsub("c","",tolower(cluster)))) %>% 
  group_by(year, clustid) %>% 
  summarise(n = n(),
            pop = unique(pop)) %>% 
  ungroup() %>% 
  mutate(cnr = n*1e5/pop) -> cluster_CNRs

compare_preds %>% 
  full_join(filter(cluster_CNRs, year == 2020)) %>%
  ggplot(aes(cnr, Estimate, 
             # xmin = .lower, xmax = .upper,
             ymin = `Q2.5`, ymax = `Q97.5`)) +
  geom_point(alpha = 0.5) +
  # geom_errorbar(alpha = 0.1) +
  # geom_errorbarh(alpha = 0.1) +
  facet_wrap(~Model) +
  scale_x_continuous(trans = "log2") +
  labs(x = "Case notification rate, 2020", y = "Testing history model: fitted prevalence") -> plot_fit_CNR

compare_iids %>% 
  full_join(filter(cluster_CNRs, year == 2020)) %>%
  ggplot(aes(cnr, mean, 
             # xmin = .lower, xmax = .upper,
             ymin = `2.5%`, ymax = `97.5%`)) +
  geom_point(alpha = 0.5) +
  # geom_errorbar(alpha = 0.1) +
  # geom_errorbarh(alpha = 0.1) +
  facet_wrap(~Model) +
  scale_x_continuous(trans = "log2") +
  labs(x = "Case notification rate, 2020", y = "Testing history model: cluster residual") -> plot_resid_CNR


plot_resid_CNR
plot_fit_CNR


```

# Missing HIV status

```{r make_data}

cluster_inc <- prev_clust %>% 
  st_drop_geometry() %>% 
  dplyr::select(cluster_number, clust_inc_1518)

# First drop 3 with missing ages
dat_nahiv <- dat %>% 
  filter(!is.na(agegp))

# Redefine HIV status
dat_nahiv$hiv <- dat_nahiv$hiv_combined
dat_nahiv$hiv[dat_nahiv$hiv == "HIV positive ART"] <- "HIV positive"

dat_nahiv <- mutate(dat_nahiv, 
              test01 = as.numeric(ever_test == "Yes"),
              agegp = factor(cut(s09age, 
                                 breaks = c(17,24,34,44,54,120)),
                        levels = c("(17,24]","(24,34]","(34,44]","(44,54]","(54,120]"),
                        labels = c("(17,24]","(24,34]","(34,44]","(44,54]","> 55")),
              hiv = factor(hiv, levels = c("HIV negative","HIV positive"), labels = c("HIV negative","HIV positive")),
              pov_score_scale = -scale(pov_score),
              wealth_quant = as.factor(ntile(pov_score_scale, 6)),
              pov01 = as.factor(pov_score > 0),
              wealth_quant1 = as.factor(wealth_quant == 1),
              wealth_step = as.factor(wealth_step),
              wealth_step1 = as.factor(wealth_step == 1),
              know_test = as.factor(s60tb_anyone == 1),
              know_trt = as.factor(s62tb_famtb == 1),
              know_died = as.factor(s65tb_died == 1)) %>% 
  full_join(cluster_inc, by = c("clustid" = "cluster_number"))

hivlabs <- c("HIV negative","HIV positive", "HIV unknown")
dat_uhiv <- dat_nahiv %>% 
  mutate(hiv = factor(replace_na(as.character(hiv), "HIV unknown"), 
         levels = hivlabs, 
         labels = hivlabs))
# Check
dat_uhiv %>% 
  dplyr::select(test01, sex, agegp, hiv, pov_score, wealth_quant, wealth_step) %>% 
  summary() %>% 
  print()

```


First simply refit with HIV unknown as a factor level. Still drop missing poverty scores at this stage for comparability. 

```{r hiv_unkown}

fit_uhiv <- update(fit_base, formula = "test01 ~ sex*agegp + hiv + (1|clustid)", 
                    newdata = filter(dat_uhiv, !is.na(pov_score)))

print(summary(fit_uhiv))
print(exp(fixef(fit_uhiv)))
```

```{r plot_hiv_fixeff}

labs <- c("Male",
          "Age (24,34]", "Age (34,44]", "Age (44,54]", "Age > 55",
          "Male : Age (24,34]","Male : Age (34,44]","Male : Age (44,54]", "Male : Age > 55",
          "HIV positive", "HIV unknown")

get_model_data(fit_uhiv, type = "est", transform = NULL) %>% 
  mutate(Model = "Unknown HIV included") %>% 
  bind_rows(
    mutate(get_model_data(fit_base, type = "est", transform = NULL),
           Model = "Unknown HIV excluded")) %>% 
  mutate(order = row_number(),
         term = factor(term, levels = unique(term)[c(1:5,8:11,6,7)])) %>% 
  ungroup() -> logORtab

pos <- position_dodge()
logORtab %>% 
  ggplot(aes(term, estimate, col = Model)) +
  geom_hline(yintercept = 0, col = "grey", lty = "dashed") +
  geom_linerange(aes(ymin = conf.low, ymax = conf.high), position = position_dodge(width = 0.5), lwd = 1) +
  geom_linerange(aes(ymin = conf.low50, ymax = conf.high50), position = position_dodge(width = 0.5), lwd = 1.2) +
  geom_point(position = position_dodge(width = 0.5), col = "white", pch = 15, cex = 0.8) +
  scale_colour_viridis_d(option = "turbo") +
  coord_flip() +
  scale_x_discrete(limits = rev(levels(logORtab$term)), labels = rev(labs)) +
  labs(x = "", y = "Estimate (log-odds ratio)") +
  theme(text = element_text(size = 14))

```

## Multiple imputation

### Method 1: Impute data before fit

```{r mi_hiv}

dat_imphiv <- filter(dat_nahiv, !is.na(pov_score)) %>%  
  dplyr::select(c(test01, clustid, sex, agegp, hiv, 
                  pov_score, wealth_step,
                  know_test, know_trt, know_died)) %>% 
  mice::mice(m = 5, print = FALSE, method = "cart")

fit_imphiv <- brm_multiple(test01 ~ sex*agegp + hiv + (1|clustid), 
                        data = dat_imphiv, 
                        chains = 4,
                        family = bernoulli(),
                        prior = priors)

print(summary(fit_imphiv))
print(exp(fixef(fit_imphiv)))
```

## Compare fixed effects

Extract coefficients from three alternative actions for unknown HIV (exclude, fit as a factor level, impute) and plot side-by-side for comparison. 

+ Multiple imputation appears to shrink all estimates to zero. 

```{r plot_fitimp}

get_model_data(fit_uhiv, type = "est", ci.lvl = 0.95, transform = NULL) %>% 
  mutate(Model = "Included as factor level") %>% 
  bind_rows(
    mutate(get_model_data(fit_base, type = "est", ci.lvl = 0.95, transform = NULL),
           Model = "Excluded")) %>% 
  bind_rows(
    mutate(get_model_data(fit_imphiv, type = "est", ci.lvl = 0.95, transform = NULL),
           Model = "Imputed")) %>% 
  mutate(order = row_number(),
         term = factor(term, levels = unique(term)[c(1:5,8:11,6,7)])) %>% 
  ungroup() -> logORtab2

pos <- position_dodge()
logORtab2 %>% 
  ggplot(aes(term, estimate, col = Model)) +
  geom_hline(yintercept = 0, col = "grey", lty = "dashed") +
  geom_linerange(aes(ymin = conf.low, ymax = conf.high), position = position_dodge(width = 0.5), lwd = 0.5) +
  geom_linerange(aes(ymin = conf.low50, ymax = conf.high50), position = position_dodge(width = 0.5), lwd = 0.8) +
  geom_point(position = position_dodge(width = 0.5), pch = 15, cex = 1.2) +
  coord_flip() +
  scale_colour_viridis_d(option = "turbo", begin = 0.1, end = 0.9) +
  scale_x_discrete(limits = rev(levels(logORtab2$term)), labels = rev(labs)) +
  labs(x = "", y = "Estimate (log-odds ratio)", col = "Unknown HIV action") +
  theme(text = element_text(size = 14)) -> comp_hiv_mods
comp_hiv_mods

ggsave(here::here("figures/manuscript/supplementary", "figs7.png"), height = 6, width = 8)

```

## Impute HIV and HH vars

```{r mi_all}

dat_impall <- dat_nahiv %>% 
  dplyr::select(c(test01, clustid, sex, agegp, hiv, 
                  wealth_quant, wealth_step,
                  know_test, know_trt, know_died)) %>% 
  mice::mice(m = 5, print = FALSE, method = "cart")

fit_impall_quant <- brm_multiple(test01 ~ sex*agegp + hiv + wealth_quant + (1|clustid), 
                        data = dat_impall, 
                        chains = 4,
                        family = bernoulli(),
                        prior = priors)

print(summary(fit_impall_quant))
print(exp(fixef(fit_impall_quant)))


fit_impall_step <- brm_multiple(test01 ~ sex*agegp + hiv + wealth_step + (1|clustid), 
                        data = dat_impall, 
                        chains = 4,
                        family = bernoulli(),
                        prior = priors)

print(summary(fit_impall_step))
print(exp(fixef(fit_impall_step)))

```

```{r plot_fitsimppov}

get_model_data(fit_list_pov$score_q, type = "est", ci.lvl = 0.95, transform = NULL) %>% 
  mutate(Model = "Complete case: poverty score") %>%   
  bind_rows(
    mutate(get_model_data(fit_list_pov$step, type = "est", ci.lvl = 0.95, transform = NULL),
           Model = "Complete case: self-assessed wealth")) %>% 
  bind_rows(
    mutate(get_model_data(fit_impall_quant, type = "est", ci.lvl = 0.95, transform = NULL),
           Model = "Imputed: poverty score")) %>% 
  bind_rows(
    mutate(get_model_data(fit_impall_step, type = "est", ci.lvl = 0.95, transform = NULL),
           Model = "Imputed: self-assessed wealth")) %>% 
  mutate(order = row_number(),
         term = factor(term, levels = unique(term)[c(1:5,12:15,6,7:11,16:20)])) %>% 
  ungroup() -> logORtab3

labs2 <- c("Male",
          "Age (24,34]", "(34,44]", "(44,54]", "> 55",
          "Male : Age (24,34]","(34,44]","(44,54]", "> 55",
          "HIV positive", 
          "PMT score quantile: 2","3","4","5", "6",
          "Self-assessed wealth step: 2", "3", "4", "5", "6")

pos <- position_dodge()
logORtab3 %>% 
  ggplot(aes(term, estimate, col = Model)) +
  geom_hline(yintercept = 0, col = "grey", lty = "dashed") +
  geom_vline(xintercept = 5.5, col = "grey") +
  geom_vline(xintercept = 10.5, col = "grey") +
  geom_linerange(aes(ymin = conf.low, ymax = conf.high), position = position_dodge(width = 0.5), lwd = 0.5) +
  geom_linerange(aes(ymin = conf.low50, ymax = conf.high50), position = position_dodge(width = 0.5), lwd = 0.8) +
  geom_point(position = position_dodge(width = 0.5), pch = 15, cex = 1.2) +
  coord_flip() +
  scale_colour_viridis_d(option = "turbo", begin = 0.1, end = 0.9) +
  scale_x_discrete(limits = rev(levels(logORtab3$term)), labels = rev(labs2)) +
  labs(x = "", y = "Estimate (log-odds ratio)", col = "") +
  theme(text = element_text(size = 14)) -> comp_imp_all
comp_imp_all

ggsave(here::here("figures/manuscript/supplementary", 
                  "figs8.png"), height = 8, width = 10)

```

```{r save_fits}

imp.fit.list <- list("unknHIV" = fit_uhiv,
                     "HIV" = fit_imphiv, 
                     "HIV_pov_quant" = fit_impall_quant,
                     "HIV_pov_step" = fit_impall_step)

saveRDS(imp.fit.list, here::here("output","fits_impute.rds"))

```
